+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 ------
 * change UI workflow for appending a session
 * Video QC: changed log level from WARNING to INFO if less than 0.1% of frames have been dropped
+* Valve: Use restricted quadratic fit when converting reward volume to valve opening time
 
 8.24.0
 ------